Majid_App/requirements.txt
streamlit
SpeechRecognition
pyttsx3
transformers
torch
python-Levenshtein
arabic-reshaper
python-bidi
sounddevice
numpy
PyAudio 